mvcc/reader/scanner: refactor ScanBuilder
